(hel-118) factorisation des graphiques des view models
