+++ src/frontend/ui/établissement-territorial-sanitaire/ÉtablissementTerritorialSanitaireViewModel.tsx
@@ -1,25 +1,10 @@
-import {
-  BarElement,
-  CategoryScale,
-  Chart as ChartJS,
-  Chart,
-  ChartData,
-  Legend,
-  LegendItem,
-  LinearScale,
-  LineController,
-  LineElement,
-  PointElement,
-  Title,
-  Tooltip,
-} from 'chart.js'
-import ChartDataLabels from 'chartjs-plugin-datalabels'
 import { ReactElement } from 'react'
 import { Bar } from 'react-chartjs-2'
 
 import { ÉtablissementTerritorialSanitaire } from '../../../backend/métier/entities/établissement-territorial-sanitaire/ÉtablissementTerritorialSanitaire'
 import { ÉtablissementTerritorialSanitaireActivité } from '../../../backend/métier/entities/établissement-territorial-sanitaire/ÉtablissementTerritorialSanitaireActivité'
 import { Wording } from '../../configuration/wording/Wording'
+import { GraphiqueViewModel } from '../commun/Graphique/GraphiqueViewModel'
 import { TableIndicateur } from '../commun/TableIndicateur/TableIndicateur'
 import styles from './BlocActivitéSanitaire.module.css'
 
@@ -32,7 +17,7 @@ type DonnéesDeDiagrammeDesSéjoursMCO = Readonly<{
   nombreSéjoursPartielsObstétrique: {x: number, y: number | null}[]
 }>
 
-export class ÉtablissementTerritorialSanitaireViewModel {
+export class ÉtablissementTerritorialSanitaireViewModel extends GraphiqueViewModel {
   readonly couleurDuFondHistogrammeBleuClair = '#DEE5FD'
   readonly couleurDuFondHistogrammeBleuFoncé = '#2F4077'
   readonly couleurDuFondHistogrammeVertClair = '#DFFDF7'
@@ -41,41 +26,9 @@ export class ÉtablissementTerritorialSanitaireViewModel {
   readonly couleurDuFondHistogrammeRougeFoncé = '#A94645'
   readonly couleurDesAxesHorizontaux = '#161616'
   readonly identifiantDeLaLégende = 'légende-graphique-sanitaire'
-  readonly couleurDuFond = '#E8EDFF'
-  readonly couleurDuFondHistogrammePrimaire = '#000091'
-  readonly couleurDuFondHistogrammeSecondaire = '#4E68BB'
-  readonly couleurDuFondDeLaLigne = '#929292'
-  readonly couleurDuFondHistogrammeDeDépassement = '#C9191E'
-  readonly couleurDelAbscisse = '#161616'
-  readonly couleurDeLaValeur = '#3A3A3A'
-  readonly fondDeCouleurPourPremierHistogramme: string[]
-  readonly fondDeCouleurPourSecondHistogramme: string[]
-  readonly grosseursDePolicePourLesLibellés: string[]
 
-  constructor(private readonly établissementTerritorial: ÉtablissementTerritorialSanitaire, private readonly wording: Wording) {
-    const nombreIndicateursActivité = établissementTerritorial.activités.length
-
-    this.fondDeCouleurPourPremierHistogramme = Array(nombreIndicateursActivité)
-      .fill(this.couleurDuFondHistogrammeSecondaire, 0, nombreIndicateursActivité - 1)
-      .fill(this.couleurDuFondHistogrammePrimaire, nombreIndicateursActivité - 1, nombreIndicateursActivité)
-    this.fondDeCouleurPourSecondHistogramme = Array(nombreIndicateursActivité).fill(this.couleurDuFond)
-    this.grosseursDePolicePourLesLibellés = Array(nombreIndicateursActivité)
-      .fill('normal', 0, nombreIndicateursActivité - 1)
-      .fill('bold', nombreIndicateursActivité - 1, nombreIndicateursActivité)
-
-    ChartJS.register(
-      BarElement,
-      CategoryScale,
-      ChartDataLabels,
-      Legend,
-      LinearScale,
-      LineController,
-      LineElement,
-      PointElement,
-      Title,
-      Tooltip,
-      this.construisLePluginDeLégende()
-    )
+  constructor(private readonly établissementTerritorial: ÉtablissementTerritorialSanitaire, wording: Wording) {
+    super(wording, établissementTerritorial.activités.length)
   }
 
   public get titre(): string {
@@ -254,100 +207,10 @@ export class ÉtablissementTerritorialSanitaireViewModel {
     )
   }
 
-  private construisLePluginDeLégende() {
-    function créeLeLibelléPourLaLégende(chart: ChartJS, libellé: LegendItem): HTMLLIElement {
-      const li = document.createElement('li')
-
-      li.onclick = () => {
-        chart.setDatasetVisibility(libellé.datasetIndex, !chart.isDatasetVisible(libellé.datasetIndex))
-        chart.update()
-      }
-
-      const boxSpan = document.createElement('span')
-      boxSpan.style.background = libellé.fillStyle as string
-      boxSpan.style.border = `solid ${libellé.strokeStyle} 1px`
-      boxSpan.innerHTML = '&nbsp;'
-
-      const textContainer = document.createElement('p')
-      textContainer.classList.add('fr-text--xs')
-      textContainer.style.textDecoration = libellé.hidden ? 'line-through' : ''
-
-      const text = document.createTextNode(libellé.text)
-      textContainer.appendChild(text)
-
-      li.appendChild(boxSpan)
-      li.appendChild(textContainer)
-      return li
-    }
-
-    return {
-      afterUpdate(chart: Chart, _args: Object, options: any) {
-        const légende = document.getElementById(options.containerID)
-
-        if (!légende)
-          return
-
-        while (légende.firstChild) {
-          légende.firstChild.remove()
-        }
-
-        // @ts-ignore
-        const libellésDeLaLégende = chart.options.plugins?.legend?.labels.generateLabels(chart) || []
-
-        libellésDeLaLégende.forEach((libellé) => {
-          const libelléDeLégende = créeLeLibelléPourLaLégende(chart, libellé)
-          légende.appendChild(libelléDeLégende)
-        })
-      },
-      id: 'htmlLegend',
-    }
-  }
-
-  private optionsHistogrammeÀBandes(idDeLaLégende: string) {
-    return {
-      animation: false,
-      elements: { bar: { borderWidth: 2 } },
-      plugins: {
-        datalabels: { display: false },
-        htmlLegend: { containerID: idDeLaLégende },
-        legend: { display: false },
-        tooltip: { callbacks: { label: this.tooltip(this.wording, this.insèreUnEspaceTousLes3Chiffres) } },
-      },
-      responsive: true,
-      scales: {
-        x: { grid: { drawOnChartArea: false }, ticks: { color: 'var(--text-default-grey)' } },
-        y: { grid: { color: this.couleurDesAxesHorizontaux, drawBorder: false }, stacked: true, ticks: { color: 'var(--text-default-grey)' } },
-      },
-    }
-  }
-
   private valeursDesNombresDeSéjours(nombresSéjours: {x: number, y: number | null}[]): (string | null)[] {
     return nombresSéjours.map((nombreSéjour) => {
       return nombreSéjour.y ? this.insèreUnEspaceTousLes3Chiffres(nombreSéjour.y) : null
     })
-  }
-
-  private tooltip(wording: Wording, formateLesNombres: Function) {
-    return function (context: any) {
-      const label = `${context.dataset.label} : ${formateLesNombres(context.parsed.y)}`
-
-      if (context.datasetIndex <= 1) {
-        const nombreSéjoursHospitalisationPartielleMédecine = context.parsed._stacks.y['0']
-        const nombreSéjoursHospitalisationComplèteMédecine = context.parsed._stacks.y['1']
-        return [label, `${wording.TOTAL_HOSPITALISATION_MÉDECINE} : ${formateLesNombres(nombreSéjoursHospitalisationPartielleMédecine + nombreSéjoursHospitalisationComplèteMédecine)}`]
-      }
-      if (context.datasetIndex === 2 || context.datasetIndex === 3) {
-        const nombreSéjoursHospitalisationPartielleChirurgie = context.parsed._stacks.y['2']
-        const nombreSéjoursHospitalisationComplèteChirurgie = context.parsed._stacks.y['3']
-        return [label, `${wording.TOTAL_HOSPITALISATION_CHIRURGIE} : ${formateLesNombres(nombreSéjoursHospitalisationPartielleChirurgie + nombreSéjoursHospitalisationComplèteChirurgie)}`]
-      }
-      if (context.datasetIndex === 4 || context.datasetIndex === 5) {
-        const nombreSéjoursHospitalisationPartielleObstétrique = context.parsed._stacks.y['4']
-        const nombreSéjoursHospitalisationComplèteObstétrique = context.parsed._stacks.y['5']
-        return [label, `${wording.TOTAL_HOSPITALISATION_OBSTÉTRIQUE} : ${formateLesNombres(nombreSéjoursHospitalisationPartielleObstétrique + nombreSéjoursHospitalisationComplèteObstétrique)}`]
-      }
-      return label
-    }
   }
 
   private formateLeTitreDeLEntitéJuridiqueDeRattachement() {
@@ -361,10 +224,6 @@ export class ÉtablissementTerritorialSanitaireViewModel {
 
   private insèreUnEspaceTousLesNCaractères(str: string, nombreDeCaractères: number): string {
     return str.split('').map((letter, index) => index % nombreDeCaractères === 0 ? ' ' + letter : letter).join('').trim()
-  }
-
-  private insèreUnEspaceTousLes3Chiffres(num: number): string {
-    return num.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ' ')
   }
 
   private formateLaDate(date: string): string {
@@ -426,81 +285,5 @@ export class ÉtablissementTerritorialSanitaireViewModel {
     })
 
     return [valeurs, années]
-  }
-
-  private afficheUnHistogrammeHorizontal(
-    chartColors: string[],
-    valeurs: number[],
-    dataLabelsColor: string[],
-    années: number[],
-    identifiant: string
-  ): JSX.Element {
-    const data: ChartData = {
-      datasets: [
-        {
-          backgroundColor: chartColors,
-          data: valeurs,
-          datalabels: { labels: { title: { color: dataLabelsColor } } },
-          maxBarThickness: 60,
-          type: 'bar',
-          yAxisID: 'y',
-        },
-      ],
-      labels: années,
-    }
-
-    return (
-      <>
-        <Bar
-          // @ts-ignore
-          data={data}
-          // @ts-ignore
-          options={this.optionsHistogrammeHorizontal(Math.max(...valeurs))}
-        />
-        <TableIndicateur
-          identifiants={[identifiant]}
-          libellés={années}
-          valeurs={[valeurs]}
-        />
-      </>
-    )
-  }
-
-  private optionsHistogrammeHorizontal(valeurMaximale: number) {
-    return {
-      animation: false,
-      indexAxis: 'y',
-      plugins: {
-        datalabels: {
-          align: 'end',
-          anchor: 'end',
-          font: {
-            family: 'Marianne',
-            size: 14,
-          },
-        },
-        legend: { display: false },
-        tooltip: { enabled: false },
-      },
-      scales: {
-        x: {
-          display: false,
-          max: 1.15 * valeurMaximale,
-          min: 0,
-        },
-        y: {
-          grid: {
-            drawBorder: false,
-            drawOnChartArea: false,
-            drawTicks: false,
-          },
-          ticks: {
-            color: this.couleurDelAbscisse,
-            // TODO
-            font: { weight: this.grosseursDePolicePourLesLibellés },
-          },
-        },
-      },
-    }
   }
 }
